- L2seg -ve cases, lport skipping 0, net -> SG (#1493)
* - Adding .ini file to bring up

* - Fixing json error

* - Adding registered mac programming

* - L2seg -ve cases, lport skipping 0, net -> SG

* - adding slab leak for test case

+++ nic/hal/pd/common/pd_api.hpp
@@ -54,7 +54,7 @@ using hal::cpucb_t;
 using hal::rawrcb_t;
 using hal::rawccb_t;
 
-typedef uint16_t    l2seg_hw_id_t;
+typedef uint32_t    l2seg_hw_id_t;
 
 typedef struct pd_tenant_args_s {
     tenant_t           *tenant;
